requirements.txt
pyqt5
python-vlc
pyinstaller
pillow
pip
setuptools
